webapp/requirements.txt
Flask==1.0.2
gunicorn==19.9.0
names==0.3.0
python-decouple==3.1